Did everything except statistics

+++ JavaSpringBackend/src/main/java/com/ftn/mas/elasticSearch/Backend/dtos/BodyGeoCodingDTO.java
@@ -3,5 +3,5 @@ package com.ftn.mas.elasticSearch.Backend.dtos;
 public class BodyGeoCodingDTO {
 	public Double longitude;
 	public Double latitude;
-	public Double radius;
+	public Integer radius;
 }
